refactor(core): ремедиация фундамента — этапы 0–6 аудита (BACKLOG #226–#231)
Аудит выявил фундаментальные промахи, блокирующие развитие (в т.ч. модуль коуча);
все этапы плана выполнены за одну сессию, db-safety-ревью — GO.

Этап 0: рантайм-фиксы — /stats бота (db.func), мёртвый reanalyze (несуществующая
колонка interval_min_phase_distance_m), performance Integer→Float (Coros отдаёт
float −2..+2, PG округлял).
Этап 1 (#226): opt-in PG-режим тестов через TEST_PG_URL (alembic-схема, строго
localhost), PRAGMA foreign_keys для SQLite, CI гоняет оба режима, §6 CLAUDE.md.
Этап 2 (#227): сбой синка → -1 (таймстемп не двигается — раньше данные терялись
навсегда), счётчики сбоев + telegram-notify на 3-м, backoff, кэш токена, throttle.
Этап 3 (#228): external_activity_id/source_brand/file_sha256 + частичные UNIQUE,
sync/dedup.py (ID-first, окно ±120с только для legacy), SHA256 для ручных загрузок,
bin/backfill_external_ids.py (--dry-run).
Этап 4 (#229): сырые FIT/TCX в uploads/raw (content-addressed), reanalyze от сырья,
volume ./uploads в bot, bin/backfill_raw_fits.py.
Этап 5 (#230): пороги — только в coach/config.py, sleep_quality удалён из весов,
анти-дрейф-тесты.
Этап 6 (#231): один канонический get_db, db параметром в сервисах/репозиториях,
SessionLocal только в композиционных корнях + тест-гвард.

Миграции: k4l5m6n7o8p9 → l5m6n7o8p9q0 → m6n7o8p9q0r1 → n7o8p9q0r1s2 (все аддитивные).
Тесты: 222 passed в обоих режимах (SQLite + PostgreSQL/alembic upgrade head).
Деплой: backup → build/up → backfill_external_ids (--dry-run → --apply) →
backfill_raw_fits --apply (история Coros API конечна — не откладывать).

+++ src/domain/models/health.py
@@ -22,7 +22,7 @@ class DailyMetrics(Base):
     tired_rate = Column(Integer, nullable=True)
     training_load = Column(Float, nullable=True)
     training_load_ratio = Column(Float, nullable=True)
-    performance = Column(Integer, nullable=True)
+    performance = Column(Float, nullable=True)  # Coros performance: float −2..+2 (не Integer — теряется градация)
     ati = Column(Float, nullable=True)
     cti = Column(Float, nullable=True)
     vo2max = Column(Float, nullable=True)
